Move standard library functions to std_fns.rs
Since it is growing, it deserves its own module, rather than bloating
interpreter.rs.
